A lot of audio cleanup. Replaced FFTW with our own "NIH" FFT, which might even be faster but most importantly it is quite short and integrates well with C++. There are some issues with it still, probably related to phase change detection, but I'll have to look into that later.
The only real drawback of not using FFTW is that now we are limited to
2^p size FFTs, where p must be a compile-time constant. The Analyzer
class (renamed from CFft) now uses FFT_P const variable for this, instead
of fftSize. FFT_N replaces fftSize.

We also now use boost::mutex and boost::scoped_lock, so the record part of
the program does not depend on SDL anymore and we got rid of our NIH
wrappers for SDL locks.


svn: r373