Renamed all config files, now config schema is schema.xml and config files are stored in performous folders under system config paths as config.xml.
Use Ctrl+Alt+S to save system config (requires write access to /etc) and
Ctrl+Alt+R to reset to factory default (rather than to system default).

Config system improved so that it doesn't leave extra files behind and that
it tracks the default values of factory and system configs separately.

Added a catcher for runtime errors in main loop. For now it just prints to
console, but eventually when Performous gets "flash message" support, they
should be displayed as such.