Some code added to songparser-sm.cc, yet not enough to work correctly.
The beginning of songparser-sm works as songparser-txt reading information
connected to the song. The notes are read in a vector of note "lines", each
line in form of another vector (size of 4) or map.
Additions to songparser.hh and song.hh needed and (most of them)
explained in the file.