implemented search 2.0 still need to change javascript website though
the old system send the whole song database to the webbrowser and
searching the database was local on the webbrowser itself, this turned
out to be very slow with 2000+ songs and you could only search on title
and artist, not the folders in which they are stored.

this new system uses performous to do the search it sends a string to /
matching songs, this makes the web-frontend more responsive with large
databases up to 10000 songs (believe me, I have them) and it allows you
to search on folder-name which is useful because in my database the
folders have the names of the tv-series the songs belong to, wich means
for me I can now search on tv-series.

I still have to implement this in the site but a quick test with curl
proves to work perfectly.

the old /api/getDatabase.json will still be used, I'm planning on
putting the data in a table to print out a list of available songs.