Change Duet logic so Performous doesn't crash anymore.
This changes the internal logic a bit: Since the Engine always wants the
same number of Analyzers as Vocals, we always use as much Analyzers as
vocals. In the case the user either selects "Normal Mode" from the Duet
menu or the song does not contain a duet in the first place, we just
assign the first selected vocal track (or the only one) to all
microphones.

TODO: The duet menu should somehow visualize when "Normal Mode" is
selected, that only the first vocal selection is relevant.