Solve a couple residual pango woes.
There were still a couple issues with the way pango handles paths to
config files when using relative references; took a while of testing
but those should be solved now.

For reference: pangorc uses CWD (as in, from wherever the launcher is
actually being run) as it’s base; not the executable path and not where
pangorc itself is either.

Additionally, deleted a superfluous copy of the app icon just sitting
there.