Use HINTS instead of PATHS for paths from pkg-config.
From the cmake documentation: "HINTS [...] should be paths computed by
system introspection, such as a hint provided by the location of another
item already found.  Hard-coded guesses should be specified with the
PATHS option."

This was inspired by a header conflict on my PC-BSD test system: The
portaudio.h from portaudio 18 was in a system path (scanned before the
path in PATHS) and the portaudio.h from portaudio 19 was in another
directory (which was returned by pkg-config). Switching this always
prefers directories from pkg-config over "random finds" of header files
and libraries with the same name.