Quick and dirty patch to prevent text overflow.
After some analyzing the code, I thought of this and seems to work to prevent text going off-screen.

Note, however… I don’t know what that loop at the end of the function is supposed to be doing but something I noticed is that as far as I could tell, it always does at most one iteration, because there are no syllables… _text.size() is always 1.