Make migrating config files more robust.
To avoid issues, I’d rather explicitly avoid overwriting anything.
In order for this to work seamlessly, it needs to be done during the bootstrapping phase of path configuring, and we can’t use clog yet at that point. So, add a boolean switch to let the next pass of this function know that we did in fact migrate the files.