Gross correction
This _kinda_ fixes the issue of the skip message appearing when one singer has lines and the other doesn’t.

Still, the overall behavior is flawed in that if, for example, one is singing a duet in single-player, or if both players are singing the same voice, it’s possible getVocalTrack() will return the wrong vocal track (I’m assuming this is the reason) and thus the skip message won’t always appear.