Add players as a member of ScreenSing
Avoids calculating it every time we need it and corrects the bug on starting a song; this was caused by an omission on my part to always set players to 1, even if no mics are connected. I incorporated this logic into the return value for the member definition, so as to avoid this in the future.