Remove SDL_SetWindowTitle and Icon
We were calling both of these before actually creating the window, which was raising an exception.
Besides that, I want to see whether SDL_SetWindowIcon is at all necessary nowadays. It only seems like it might be necessary in Windows and that implementation has a pretty nasty issue, it loads the image data directly on the stack, instead of using malloc.