Make character detection more robust as follows:
If we have a UTF-8 BOM, it has to be UTF-8.
In the case that we don’t… there are several possibilities:

First, if we haven’t specified a filename to the function then we’re most certainly processing text that was originated in the source or has already been processed before. In these cases, we’ll assume UTF-8.

If ICU is not sure what we’re dealing with, it’s most likely because we have only english characters. This is most likely a false-negative for UTF-8 (if there were any non-UTF-8 characters, confidence would be above 50 AFAIK), so we treat it as UTF-8.
I decided to use 75 for confidence in these cases to make it easier for ourselves to debug in the future if the previous assumption is incorrect.

OTOH, if confidence is over 50 from the get-go, we’re most likely dealing with whatever charset ICU thinks we are. If it’s UTF-8, do nothing (we’ve already removed the BOM earlier). If it’s another charset, convert it to UTF-8.
If after all this we still don’t have a definite charset (I don’t see this ever happening but who knows), filter all non-ascii characters.