Don't consider notes overlapping if they only touch each-other. This fixes getBPM crash that occurred with FoF songs.
Note: It should be verified that no logic relies on touchy notes being detected as overlapping.
Also the code in SongParser::finalize should probably be rewritten so that it doesn't rely on BPM data (which doesn't exist for FoF/MIDI format), or alternatively extract the BPM data from midiparser into SongParser.
However, until that is fixed, simply changing the comparison avoids the issue with FoF files (which never have overlapping lyrics).