Kill index OOB errors with fire.
Hopefully this is the last instance of this bug I introduced in my duet branch a while ago. Fucking thing is as insidious as it gets. I’m still not quite sure what is causing it (maybe a race condition resulting from the loops?) but this should finally get rid of it I hope.

Also made the error more verbose in case I again didn’t get it.