Prevent crash when parent path is empty()
Actually, when in root "/" directory, in the path selection tool, is
going up, an exception is sent (and not catched) by boost.

This is because parent of "/" is "" in boost, and there cannot be
iterators on "".

Note that this bug does not occurs with std::filsystem where parent of
"/" is "/" as expected.

That said, I added a check because:
* there may be other way of getting empty path, even with std::fs
* going up when in "/" is quite useless, so user do not need to see that
  "/../" is "/"