Fix another potential overflow.
This one was trickier… it’s related to the code for dealing with negative timestamps. When skipping negative timestamps, sometimes we’d get a timestamp where the last digit was odd, which in turn could result in m_data() having an odd size, and that would then trigger an overflow because the buffering code assumes two channels.