Rework path list to keep select at top level
When I added the sort on menu entries, I did not pay attention to the
fact that the "add this directory" was now at the bottom of the page,
making it difficult to find.

This pathc reworks that, a remove the whole menu enties sort and replac
it by only path sort, so that the first entries that are pushed remains
at the top.