Add fixed scaling of the note graph
Actually the note graph is scaled dynamically depending on the
notes that are displayed during the next 8 seconds.
This commit adds a new configuration to switch between the dynamic
scaling and fixed scaling. Fixed scaling uses all notes of the song
so that the note graph does not change while playing.