Better way to handle content-types for HandleFile
RequestHandler: some renames, small clean-up
Clean the Request-Handler up in the destructor

Removedt the static if-else blocks; instead added a json file with a bunch of extensions we serve (orr might conceivably serve someday) and their associated content-type. The idea, is that it’s easy for someone to modify the filetypes their webserver knows about.

Also, prevent directory-listings slipping through.