Do not allow PlayerId to be optional
Having an Id to bear the optional semantic make its usage awkward
because one cannot tell if a Player is correct unless it tests its
PayerId value.

Having to handle invalid object is always painful, thus here I kept
optional in places where it makes sense (when the id *may* not be
valid).