Rename Song::LoadStatus::ERROR, far and near.
Windows macros strike again!

This error in particular is pretty insidious and not very easy to diagnose.

Instead of dumb work-arounds, I’m opting to just change the name of these to avoid collisions with macros.

- `Song::LoadStatus::ERROR` —> `Song::LoadStatus::PARSERERROR`.
- `far —> `farDistance`
- `near` —> `nearDistance`

That say, we don’t need to worry about if or when a third-party library does `#include <Windows.h>`