Nope. Universal builds are not gonna happen.
It appears that it’s almost impossible to build a universal binary for gobject-introspectable libraries because g-ir-scanner needs to run host code to generate the typelibs. It might be possible at some point on arm64 but there’s also currently a bug building elsewhere and I have no way to debug the building on intel.

Signed-off-by: Gregorio Litenstein <g.litenstein@gmail.com>